requirements.txt
streamlit
google-generativeai>=0.3.1
PyMuPDF
Pillow
python-dotenv
PyPDF2
pandas
matplotlib
seaborn
